Rename folder 'models' to types
